fix errors from simplification
